requirements.txt
python-telegram-bot
Faker